Converting A.N. pseudonym to PatientID
My queries use A.N.'s pseudonym throughout but CPR will use the LTHT PatientID to convert to our pseudonym. This update removes my previous half-hearted attempt at converting the identifiers in each script and adds the a query that provides the LookUp query for CJ to convert between A.N.'s pseudonym and LTHT PatientIDs.